requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/test.txt requirements/base.in requirements/test.in
#
atomicwrites==1.1.5       # via pytest
attrs==18.1.0             # via pytest
coverage==4.5.1           # via pytest-cov
ddt==1.1.3
factory-boy==2.8.1
faker==0.8.15             # via factory-boy
funcsigs==1.0.2           # via mock, pytest
futures==3.1.1
ipaddress==1.0.22         # via faker
mock==2.0.0
more-itertools==4.1.0     # via pytest
pbr==4.0.3                # via mock
pluggy==0.6.0             # via pytest
py==1.5.3                 # via pytest
pytest-cov==2.5.1
pytest-django==3.2.1
pytest==3.6.0             # via pytest-cov, pytest-django
python-dateutil==2.7.3    # via faker
pytz==2018.4              # via django
six==1.11.0               # via faker, mock, more-itertools, pytest, python-dateutil
text-unidecode==1.2       # via faker
